Optimized where clause for search cursor at step 6.5
Used the join function instead of multiple self-concatenations in order to build the where clause passed to the search cursor in step 6.5. This is much more efficient since strings are immutable. This also makes the code more readable and less verbose. Tested on Luxembourg, it makes step 6.5 processing time go from 25 seconds to 17 seconds.