Added a check to prevent null geometries.
Some null geometries seem to be generated. They appear at step 6.5 even though the content being read by the search cursor is actually populated at step 5. My take is that the if geometry is smaller then the feature class tolerance or resolution a null geometry replaces the geometry written in step 5, and is read at step 6.5. The code change allow the code to display a warning with the problematic way_id and keep working instead of crashing.