Replace the use of arcpy.Copy by arcpy.FeatureClassToFeatureClass
Another way of exporting the joined way and way_areas in Step 7. Convenient because it prevents the duplication of the field way_id, and prevent from appending the original feature class name in front of each field name in the output feature class. However this makes step 7 10 % slower in spite of writing less data to the disk.